refactor: confirm/page.tsxをServer Component化
ConfirmContentクライアントコンポーネントを分離し、page.tsxから"use client"を削除。
型定義含む177行の全ロジックをConfirmContentに移動。metadataエクスポートを追加。
